// frontend/lib/api/generated/model/laborItemCreateRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Orbis API
 * Orbis API
 * OpenAPI spec version: 1.0.0
 */
import type { LaborItemCreateRequestLaborType } from './laborItemCreateRequestLaborType';

export interface LaborItemCreateRequest {
  laborType?: LaborItemCreateRequestLaborType;
  unitPrice?: number;
  manMonth?: number;
  supplyPrice?: number;
}