// frontend/lib/api/generated/model/quotationCreateRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Orbis API
 * Orbis API
 * OpenAPI spec version: 1.0.0
 */
import type { LaborItemCreateRequest } from './laborItemCreateRequest';
import type { SolutionItemCreateRequest } from './solutionItemCreateRequest';

export interface QuotationCreateRequest {
  projectOpportunityId?: number;
  quotationDate?: string;
  paymentCondition?: string;
  note?: string;
  quotationSolutionItems?: SolutionItemCreateRequest[];
  quotationLaborItems?: LaborItemCreateRequest[];
}