// frontend/lib/api/generated/model/quotationResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Orbis API
 * Orbis API
 * OpenAPI spec version: 1.0.0
 */
import type { LaborItemResponse } from './laborItemResponse';
import type { SolutionItemResponse } from './solutionItemResponse';

export interface QuotationResponse {
  id?: number;
  quotationCode?: string;
  projectOpportunityId?: number;
  quotationDate?: string;
  paymentCondition?: string;
  consumerTotalPrice?: number;
  supplyTotalPrice?: number;
  laborTotalPrice?: number;
  totalPrice?: number;
  note?: string;
  quotationSolutionItems?: SolutionItemResponse[];
  quotationLaborItems?: LaborItemResponse[];
}